Add missing shim scalatest suites
- remove buildver from class name for easier scripting, it's already
  present in the package name
