+++ sql-plugin/src/test/311/scala/com/nvidia/spark/rapids/shims/spark311/SparkShimsSuite.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, NVIDIA CORPORATION.
+ * Copyright (c) 2023, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,20 +14,19 @@
  * limitations under the License.
  */
 
-package com.nvidia.spark.rapids.shims.spark340;
+package com.nvidia.spark.rapids.shims.spark311;
 
 import com.nvidia.spark.rapids.{ShimLoader, SparkShimVersion}
 import com.nvidia.spark.rapids.shims.SparkShimImpl
 import org.scalatest.FunSuite
 
-class Spark340ShimsSuite extends FunSuite {
+class SparkShimsSuite extends FunSuite {
   test("spark shims version") {
-    assert(SparkShimImpl.getSparkShimVersion === SparkShimVersion(3, 4, 0))
+    assert(SparkShimImpl.getSparkShimVersion === SparkShimVersion(3, 1, 1))
   }
 
   test("shuffle manager class") {
     assert(ShimLoader.getRapidsShuffleManagerClass ===
-      classOf[com.nvidia.spark.rapids.spark340.RapidsShuffleManager].getCanonicalName)
+      classOf[com.nvidia.spark.rapids.spark311.RapidsShuffleManager].getCanonicalName)
   }
-
 }
