+++ sql-plugin/src/test/320/scala/com/nvidia/spark/rapids/shims/spark320/SparkShimsSuite.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021-2022, NVIDIA CORPORATION.
+ * Copyright (c) 2021-2023, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.nvidia.spark.rapids.shims.spark322;
+package com.nvidia.spark.rapids.shims.spark320;
 
 import com.nvidia.spark.rapids.{ShimLoader, SparkShimVersion, TypeSig}
 import com.nvidia.spark.rapids.shims.SparkShimImpl
@@ -22,17 +22,17 @@ import org.scalatest.FunSuite
 
 import org.apache.spark.sql.types.{DayTimeIntervalType, YearMonthIntervalType}
 
-class Spark322ShimsSuite extends FunSuite {
+class SparkShimsSuite extends FunSuite {
   test("spark shims version") {
-    assert(SparkShimImpl.getSparkShimVersion === SparkShimVersion(3, 2, 2))
+    assert(SparkShimImpl.getSparkShimVersion === SparkShimVersion(3, 2, 0))
   }
 
   test("shuffle manager class") {
     assert(ShimLoader.getRapidsShuffleManagerClass ===
-      classOf[com.nvidia.spark.rapids.spark322.RapidsShuffleManager].getCanonicalName)
+      classOf[com.nvidia.spark.rapids.spark320.RapidsShuffleManager].getCanonicalName)
   }
 
-  test("TypeSig322") {
+  test("TypeSig") {
     val check = TypeSig.DAYTIME + TypeSig.YEARMONTH
     assert(check.isSupportedByPlugin(DayTimeIntervalType()) == true)
     assert(check.isSupportedByPlugin(YearMonthIntervalType()) == true)
