+++ sql-plugin/src/test/330cdh/scala/com/nvidia/spark/rapids/shims/spark330cdh/SparkShimsSuite.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, NVIDIA CORPORATION.
+ * Copyright (c) 2023, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,25 +14,24 @@
  * limitations under the License.
  */
 
-package com.nvidia.spark.rapids.shims.spark323;
+package com.nvidia.spark.rapids.shims.spark330cdh;
 
-import com.nvidia.spark.rapids.{ShimLoader, SparkShimVersion, TypeSig}
-import com.nvidia.spark.rapids.shims.SparkShimImpl
+import com.nvidia.spark.rapids._
 import org.scalatest.FunSuite
 
 import org.apache.spark.sql.types.{DayTimeIntervalType, YearMonthIntervalType}
 
-class Spark323ShimsSuite extends FunSuite {
+class SparkShimsSuite extends FunSuite {
   test("spark shims version") {
-    assert(SparkShimImpl.getSparkShimVersion === SparkShimVersion(3, 2, 3))
+    assert(VersionUtils.cmpSparkVersion(3, 3, 0) === 0)
   }
 
   test("shuffle manager class") {
     assert(ShimLoader.getRapidsShuffleManagerClass ===
-      classOf[com.nvidia.spark.rapids.spark323.RapidsShuffleManager].getCanonicalName)
+      classOf[com.nvidia.spark.rapids.spark330cdh.RapidsShuffleManager].getCanonicalName)
   }
 
-  test("TypeSig323") {
+  test("TypeSig") {
     val check = TypeSig.DAYTIME + TypeSig.YEARMONTH
     assert(check.isSupportedByPlugin(DayTimeIntervalType()) == true)
     assert(check.isSupportedByPlugin(YearMonthIntervalType()) == true)
